Move HandlePtr to dedicated header
