+++ Source/SoundInterface.h
@@ -29,19 +29,9 @@
 #include <memory>
 #include <wrl/client.h>
 #include "str_conv/str_conv.hpp"
+#include "utils/handle_ptr.h"
 
 using Microsoft::WRL::ComPtr;
-
-struct CloseHandleT {
-	void operator()(HANDLE handle) {
-		if (handle) {
-			CloseHandle(handle);
-		}
-	}
-};
-
-/// HANDLE is void*, so unique_ptr<void, CloseHandleT> is a RAII HANDLE.
-using HandlePtr = std::unique_ptr<void, CloseHandleT>;
 
 // Return values from CSoundStream::WaitForReady()
 enum class WaitResult {
